Streamline devcontainer setup
Use fewer files & directories mapped into the container, and instead
rely more on an .env directory that contains keys. Use environment
variables for tools where it is easier than default files.

Move to using a bashrc from another workspace since that is project
specific and can make setup of the environment tied up with
management of that environment (e.g. ansible inventory can be used
in that project to setup this environment)